solved problem with report pet phone number

+++ server/models/report.ts
@@ -4,10 +4,10 @@ import { sequelize } from "./conn";
 export class Report extends Model {}
 
 Report.init(
-    {
-        fullName: DataTypes.STRING,
-        phoneNumber: DataTypes.INTEGER,
-        description: DataTypes.STRING,
-    },
-    { sequelize, modelName: "report" }
+  {
+    fullName: DataTypes.STRING,
+    phoneNumber: DataTypes.BIGINT,
+    description: DataTypes.STRING,
+  },
+  { sequelize, modelName: "report" }
 );
